# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: CipherQuest API
  description: |
    A comprehensive cybersecurity learning platform API that provides authentication, 
    module management, CTF challenges, leaderboards, and user progress tracking.
    
    ## Features
    - JWT-based authentication with OAuth support (Google, GitHub)
    - Learning modules with progress tracking
    - CTF challenges with flag submission
    - Real-time leaderboards
    - User profile and progress management
    - Admin dashboard and content management
    
    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <your_access_token>
    ```
    
    ## Rate Limiting
    - Auth endpoints: 5-10 requests per minute
    - Challenge submission: 20 requests per minute
    - Other endpoints: 30-60 requests per minute
  version: 1.0.0
  contact:
    name: CipherQuest Support
    email: support@cipherquest.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cipherquest.com
    description: Production server
  - url: https://cipherquest-api.herokuapp.com
    description: Heroku deployment
  - url: http://localhost:5000
    description: Local development server
security:
  - BearerAuth: []
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9_]+$'
                  example: "hacker123"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  minLength: 8
                  example: "securePassword123"
                first_name:
                  type: string
                  maxLength: 50
                  example: "John"
                last_name:
                  type: string
                  maxLength: 50
                  example: "Doe"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  user:
                    $ref: '#/components/schemas/User'
                  access_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                  refresh_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user with username/email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                username:
                  type: string
                  example: "hacker123"
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "securePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  user:
                    $ref: '#/components/schemas/User'
                  access_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                  refresh_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Logout current user (token invalidation)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout successful"
  /api/auth/google:
    post:
      tags:
        - Authentication
      summary: Google OAuth login
      description: Authenticate user with Google OAuth token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - access_token
              properties:
                access_token:
                  type: string
                  description: Google OAuth access token
                  example: "ya29.a0AfH6SMC..."
      responses:
        '200':
          description: OAuth login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OAuth login successful"
                  user:
                    $ref: '#/components/schemas/User'
                  access_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                  refresh_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
  /api/auth/github:
    post:
      tags:
        - Authentication
      summary: GitHub OAuth login
      description: Authenticate user with GitHub OAuth token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - access_token
              properties:
                access_token:
                  type: string
                  description: GitHub OAuth access token
                  example: "gho_..."
      responses:
        '200':
          description: OAuth login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OAuth login successful"
                  user:
                    $ref: '#/components/schemas/User'
                  access_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                  refresh_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get current authenticated user's profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "hacker123"
        email:
          type: string
          format: email
          example: "user@example.com"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        avatar_url:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/avatar.jpg"
        experience:
          type: integer
          example: 2500
        level:
          type: integer
          example: 5
        rank:
          type: string
          example: "Elite Hacker"
        is_active:
          type: boolean
          example: true
        email_verified:
          type: boolean
          example: true
        oauth_provider:
          type: string
          nullable: true
          example: "google"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        last_login:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-20T14:45:00Z"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid credentials"
        message:
          type: string
          example: "The provided credentials are incorrect"
        code:
          type: string
          example: "AUTH_001"
        details:
          type: object
          additionalProperties: true
...